refactor ci to handle new minor releases better

+++ ci/travis-ci.sh
@@ -123,21 +123,22 @@ after-success() {
 
         # PUSH BACK TO OUR GIT REPO
         # Bump our things and reset tags
-        grunt bump-patch
+        # Dont bump or push if this is a new minor version
+        if [ "${DISCO_ARRAY[2]}" -gt "0" ]; then
+          grunt bump-patch
 
-        # Reset upstream and tags so we can push our changes to it
-        # We need to re-add this in because our clone was originally read-only
-        git remote rm origin
-        git remote add origin git@github.com:$TRAVIS_REPO_SLUG.git
-        git checkout $TRAVIS_BRANCH
-        git tag -d $DISCO_TAG
-        git push origin :$DISCO_TAG
+          # Reset upstream and tags so we can push our changes to it
+          # We need to re-add this in because our clone was originally read-only
+          git remote rm origin
+          git remote add origin git@github.com:$TRAVIS_REPO_SLUG.git
+          git checkout $TRAVIS_BRANCH
 
-        # Add all our new code and push reset tag with ci skipping on
-        git add --all
-        git commit -m "${COMMIT_MSG} VERSION ${DISCO_TAG} [ci skip]" --author="Kala C. Bot <kalacommitbot@kalamuna.com>" --no-verify
-        git tag $DISCO_TAG
-        git push origin $TRAVIS_BRANCH --tags
+          # Add all our new code and push reset tag with ci skipping on
+          git add --all
+          git commit -m "${COMMIT_MSG} VERSION ${DISCO_TAG} [ci skip]" --author="Kala C. Bot <kalacommitbot@kalamuna.com>" --no-verify
+          git tag $DISCO_TAG
+          git push origin $TRAVIS_BRANCH --tags
+        fi
 
         # NODE PACKAGES
         # Deploy to NPM
